LIBITD-1940. Made storageprovider/storagepath fields optional in CSV
Made the "storageprovider" and "storagepath" fields optional in
the input inventory CSV file. This is necessary because these fields
are not present in the CSV file generated by the "preserve" tool.

Modified the code to set the "storage_provider" and "storage_location"
fields to None if the CVS fields are not present.

https://issues.umd.edu/browse/LIBITD-1940