LIBITD-1754. Updated Tests to Run Against Postgres and Sqlite
Updated the tests significantly to run against both Postgres and Sqlite.
This involed removing the usage of the Unittest module, which hindered
the types of Pytest features that could be used. Also updated the
way assertions were called since the previous way was done which relied
on unittest. Also moved the setup and teardown methods outside the
classes.

https://issues.umd.edu/browse/LIBITD-1754