LIBITD-2221. Fixed bug in test_export
Fixed a bug by adding a commit between the load call and
the export call. The reason being is that export queries the actual
database while the load queries the session using sqlalchemy, so the
changes in the session needed to be pushed to the database or else
export wouldn't be able to see the changes from load.

https://issues.umd.edu/browse/LIBITD-2221