LIBITD-2253: Added additional parameter for time range
Added another parameter to the process function to be able
to take the created_at__lteq/created_at__gteq parameters for
specifying time range for bags to check. Updated the core and
command classes to include these as arguments, updated the main
function to catch a new exception regarding the time arguments,
and updated the tests to test these commands.

https://umd-dit.atlassian.net/browse/LIBITD-2253