LIBITD-2270: Had to change a few tests because of logging
Because the command files now use logging and don't print to stderr,
changed the tests so that they accessed the log files from pytest.
Also, realized that pytest fulfilled the features that the patch from
unittest was being used for, so removed it entirely and replaced it
with pytest equivalent. Also removed modules that weren't being used in
the test as a result.

https://umd-dit.atlassian.net/browse/LIBITD-2270