LIBITD-2289: Reorganizing log information
Adding new log information about what bags are being grabbed, commenting
out the logs for showing files already in PATSy, and changing some
information that was previously warnings to info and moving to the end
of the file. Also removed some debug informatin to show the least four
parts of the API key and API user.

https://umd-dit.atlassian.net/browse/LIBITD-2289