LIBITD-2254. Added baseline Alembic migration
Added a baseline Alembic migration that creates the necessary tables,
indices, and views for an empty database.

This enables Alembic to replace the PATSy "schema" command, and provides
the base for future database migrations.

The migration has been set up to work, even if the tables already exist.

Views are created using Alembic's "ReplaceableObjects", see
https://alembic.sqlalchemy.org/en/latest/cookbook.html#replaceable-objects

https://umd-dit.atlassian.net/browse/LIBITD-2254