LIBITD-2254. Removed PATSy "schema" command
Removed the PATSy "schema" command, as it is no longer needed, and
could interfere with the Alembic database migration tool.

Updated the documentation to remove the "schema" command, and add
information about Alembic.

Added the "pytest-alembic" to the "test" requirements, to enable the
tests to run the Alembic migrations to set up the database.

Moved the "addr" pytest fixture into "conftest.py", as that file seems
to be the expected location for fixtures shared across tests, and
removed the fixture configuration from the individual tests. Also added
a "db_gateway" fixture that use the "alembic_runner" functionality from
"pytest-alembic" to set up the database, based on the Alembic
migrations.

Moved "_compile_drop_table" method (needed for Postgres) to the
"test/__init__.py" file, so that it would be available for all the tests

Added "db_gateway" fixture to "conftest.py" to set up the database
using Alembic migrations

Updated the tests:

* Removed the "addr" fixtures configuration as it has been moved to
  to "conftest.py"
* Removed "_compile_drop_table" method, as it was been moved to
  "tests/__init__.py"
* Replaced "addr" fixture with "db_gateway" fixture

Added "tearDown" method which clears the database after each test.

https://umd-dit.atlassian.net/browse/LIBITD-2254