LIBITD-2254. "load" and "export" fail on unexpected CSV fields
Modified the "load" and "export" commands to fail if the CSV input
or output has unexpected CSV fields. The main intention is to ensure that
older CVS files that have lower-case fields in their CSV header lines
will be flagged as having an error, instead of being processed with the
lower-case fields silently ignored.

Updated the "fixture" files, to make the CSV fields in the header line
upper-case to conform with the expected format.

https://umd-dit.atlassian.net/browse/LIBITD-2254